SqueezeDSP 0.1.31
Fixed issue with gapless audio tail not being loaded.
Optimised tail process so channels processed in parallel
Added caching of calculated filters to speed up process initialisation as this was breaking gapless.
cache checks that sample rate and channels are the same and that the filters have not changed since the cache was created.